requirements.txt
-e git://github.com/simoncrowe/bot-scheduler.git#egg=bot-scheduler
-e git://github.com/realsirjoe/instagram-scraper.git@388d89b447c6ccba2dbcb43e73deff5f8fc51664#egg=igramscraper

beautifulsoup4==4.9.0
click==7.0
factory-boy==2.12.0
matplotlib==3.1.1
marshmallow==3.3.0
networkx==2.3
pandas==1.1.0
Pillow==7.1.2
pytest==6.0.1
PyYAML==5.1.2
requests==2.23.0
scikit-learn==0.23.1
